Split subchannel check from CD dump method.
